% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_benford.R
\name{calcul_p_value}
\alias{calcul_p_value}
\title{Calcul de la p-value
Une fonction pour calculer la p-value}
\usage{
calcul_p_value(
  taille,
  nb_repetition,
  quantile_vector,
  K = 4,
  probabilite_theorique,
  probabilite_empirique,
  base_smooth,
  support_vector = c(1:9)
)
}
\arguments{
\item{taille}{taille de l echantillon}

\item{nb_repetition}{nombre de repetititons}

\item{quantile_vector}{vecteur quantile}

\item{K}{nombre  de smooth test}

\item{probabilite_theorique}{vecteur de probabilite theorique}

\item{probabilite_empirique}{vecteur de probabilite empirique}

\item{base_smooth}{Base}

\item{support_vector}{vecteur numerique}
}
\value{
la fonction renvoie un vecteur de p-value
}
\description{
Calcul de la p-value
Une fonction pour calculer la p-value
}
