% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{calcul_permutations_tk}
\alias{calcul_permutations_tk}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
calcul_permutations_tk(uj.vector, K)
}
\arguments{
\item{uj.vector}{vecteur de uj}

\item{K}{numerique entre 1 et 7}
}
\value{
la fonction renvoie le polynome
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
