% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{calcul_permutations_uj}
\alias{calcul_permutations_uj}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
calcul_permutations_uj(data, K)
}
\arguments{
\item{data}{vecteur numerique}

\item{K}{numerique entre 1 et 7}
}
\value{
la fonction renvoie le polynome
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
