% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_benford.R
\name{calcul_quantile_monte_carlo}
\alias{calcul_quantile_monte_carlo}
\title{Calcul de la table de quantile Smooth test par Monte carlo
Une fonction pour calculer la table de quantile à la Benford}
\usage{
calcul_quantile_monte_carlo(
  taille,
  nb_repetition,
  digits = 1,
  K = 5,
  support_vector = c(1:9),
  base_smooth = 10
)
}
\arguments{
\item{taille}{taille de l echantillon}

\item{nb_repetition}{nombre de repetititons}

\item{digits}{FSD}

\item{K}{nombre  de smooth test}

\item{support_vector}{vecteur numerique}

\item{base_smooth}{base}
}
\value{
la fonction renvoie une matrice
}
\description{
Calcul de la table de quantile Smooth test par Monte carlo
Une fonction pour calculer la table de quantile à la Benford
}
