% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{calcul_tk}
\alias{calcul_tk}
\title{Calcul des tk
Une fonction pour calculer les tk}
\usage{
calcul_tk(data, K)
}
\arguments{
\item{data}{vecteur numérique}

\item{K}{numerique entre 1 et 7}
}
\value{
la fonction renvoie le tk
}
\description{
Calcul des tk
Une fonction pour calculer les tk
}
