% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{calcul_tk_widehat}
\alias{calcul_tk_widehat}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
calcul_tk_widehat(Tki, K, taille, base)
}
\arguments{
\item{Tki}{vecteur de tk}

\item{K}{numerique entre 1 et 7}

\item{taille}{taille de l'echantillon}

\item{base}{base logarithmique}
}
\value{
la fonction renvoie le meilleur tk
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
