% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{calcul_U_for_j}
\alias{calcul_U_for_j}
\title{Calcul des Uj
Une fonction pour calculer les Uj}
\usage{
calcul_U_for_j(j, data, taille, polynome.h.j)
}
\arguments{
\item{j}{numerique entre 1 et 7}

\item{data}{vecteur numérique}

\item{taille}{numerique}

\item{polynome.h.j}{venant la fonction h_for_j}
}
\value{
la fonction renvoie le polynome
}
\description{
Calcul des Uj
Une fonction pour calculer les Uj
}
