% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{generer_data_benford}
\alias{generer_data_benford}
\title{Generer des donnees suivant benford
Une fonction pour generer de la data suivant benford}
\usage{
generer_data_benford(n, support_vector_to_use, digits = 1, base = 10)
}
\arguments{
\item{n}{taille de l echantillon}

\item{support_vector_to_use}{vecteur numerique}

\item{digits}{FSD}

\item{base}{Base}
}
\value{
la fonction renvoie un vecteur numerique
}
\description{
Generer des donnees suivant benford
Une fonction pour generer de la data suivant benford
}
