% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{generer_probabilite_empirique}
\alias{generer_probabilite_empirique}
\title{Generer le vecteur de probabilite emprique
Une fonction pour generer le vecteur de probabilite emprique}
\usage{
generer_probabilite_empirique(data, support_vector_to_use, digits = 1)
}
\arguments{
\item{data}{vecteur numerique}

\item{support_vector_to_use}{vecteur numerique}

\item{digits}{FSD}
}
\value{
la fonction renvoie un vecteur de probabilite
}
\description{
Generer le vecteur de probabilite emprique
Une fonction pour generer le vecteur de probabilite emprique
}
