% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{generer_probabilite_theorique}
\alias{generer_probabilite_theorique}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
generer_probabilite_theorique(digit = 1)
}
\arguments{
\item{digit}{FSD}
}
\value{
la probabilite theorique de bendford
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
