% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{generer_vector_number_of_digit}
\alias{generer_vector_number_of_digit}
\title{Generer le vecteur de nombre correspondant au digit
Une fonction pour generer le vecteur de nombre correspondant au digit}
\usage{
generer_vector_number_of_digit(digits = 1)
}
\arguments{
\item{digits}{FSD}
}
\value{
la fonction un vecteur de nombre correspondant au digit
}
\description{
Generer le vecteur de nombre correspondant au digit
Une fonction pour generer le vecteur de nombre correspondant au digit
}
