% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{get_polynome_h_for_j}
\alias{get_polynome_h_for_j}
\title{Polynome Hj
Une fonction pour appliquer le smooth test à des données}
\usage{
get_polynome_h_for_j(j)
}
\arguments{
\item{j}{numerique entre 1 et 7}
}
\value{
la fonction renvoie le polynome
}
\description{
Polynome Hj
Une fonction pour appliquer le smooth test à des données
}
