% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{get_probabilite_benford_for_digit_support}
\alias{get_probabilite_benford_for_digit_support}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
get_probabilite_benford_for_digit_support(digit = 1, base = 10)
}
\arguments{
\item{digit}{FSD}

\item{base}{base logarithmique}
}
\value{
la fonction renvoie le polynome
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
