% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{get_probabilite_conjointe_n_number}
\alias{get_probabilite_conjointe_n_number}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
get_probabilite_conjointe_n_number(data, max.digits = 1, base = 10)
}
\arguments{
\item{data}{vecteur numerique}

\item{max.digits}{FSD}

\item{base}{base logarithmique}
}
\value{
la fonction renvoie le polynome
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
