% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{ksmooth.test}
\alias{ksmooth.test}
\title{Tk du test de Benford
Une fonction pour appliquer le smooth test à des données}
\usage{
ksmooth.test(data, K, digits = 1, optimal.t.permutations = FALSE)
}
\arguments{
\item{data}{un vecteur numerique}

\item{K}{un nombre entre 1 et 7}

\item{digits}{le digit a testé}

\item{optimal.t.permutations}{boolean}
}
\value{
la fonction renvoie un vecteur contenant les tk
}
\description{
Tk du test de Benford
Une fonction pour appliquer le smooth test à des données
}
