% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{BenfordSmooth.test}
\alias{BenfordSmooth.test}
\title{Tk du test de Benford
Une fonction pour appliquer le smooth test à des données}
\usage{
BenfordSmooth.test(
  data,
  K,
  digits = 1,
  alpha = 0.05,
  nb_repetition = 5000,
  optimal_t_k = TRUE
)
}
\arguments{
\item{data}{un vecteur numerique}

\item{K}{un nombre entre 1 et 7}

\item{digits}{le digit a testé}

\item{alpha}{niveau de confiance}

\item{nb_repetition}{Monte carlo repetition}

\item{optimal_t_k}{boolean}
}
\value{
la fonction renvoie un vecteur contenant les tk
}
\description{
Tk du test de Benford
Une fonction pour appliquer le smooth test à des données
}
