% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_test.R
\name{BenfordSmooth.test}
\alias{BenfordSmooth.test}
\title{Tk du test de Benford
Une fonction pour appliquer le smooth test à des données}
\usage{
BenfordSmooth.test(
  data,
  K,
  digits = 1,
  level = 0.05,
  MC_replication = 5000,
  optimal_t_k = TRUE
)
}
\arguments{
\item{data}{un vecteur numerique}

\item{K}{number of tk between 1 and 7}

\item{digits}{the digit}

\item{level}{niveau de confiance}

\item{MC_replication}{Monte carlo repetition}

\item{optimal_t_k}{boolean}
}
\value{
la fonction renvoie un vecteur contenant les tk
}
\description{
Tk du test de Benford
Une fonction pour appliquer le smooth test à des données
}
