% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilite_benford.R
\name{get_n_element_des_nombres}
\alias{get_n_element_des_nombres}
\title{Calcul du tk chapeau
Une fonction pour calculer le tk chapeau}
\usage{
get_n_element_des_nombres(data, digits = 1)
}
\arguments{
\item{data}{vecteur numerique}

\item{digits}{FSD}
}
\value{
la fonction renvoie le polynome
}
\description{
Calcul du tk chapeau
Une fonction pour calculer le tk chapeau
}
